feat: Add multiple length support
Add support for different lengths of solutions, other than just 5
letters.

This is a hidden feature (for now?) because I want to "soft-launch" it
until I decide how it should look and act. For now, it's implemented as
different languages: `nb-no-3`, `nn-no-7`, etc. The 5-letter version is
still left at `nb-no` and `nn-no` for consistency.